demo refactor of Node.js post-processors
The existing post-processors operate on YAML files line by line. That works, but it requires a lot of code that fusses with whitespace, line endings, state tracking, and so on. The resulting code isn't always easy to understand.

This demo shows how the post-processors could operate on parsed YAML objects instead of raw YAML files, resulting in simpler, clearer code. The demo includes a rewrite of the existing `deleteBaseClasses.js` post-processor, as well as a partial sketch of how `generate-devsite.sh` could be converted from shell to JS.